refactor: add SampleManager, move sample logic
